:wrench: Clean up utilities module
Clean up utilities module, removing unused code/code that has been
migrated to AnisotropIO.
